pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
color = true
line-length = 88
skip-string-normalization = true
target-version = ['py311']

[tool.isort]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = 'pydantic'
line_length = 88
multi_line_output = 3

[tool.mypy]
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_expr = false
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
explicit_package_bases = true
follow_imports = 'silent'
namespace_packages = true
no_implicit_optional = true
no_implicit_reexport = true
python_version = '3.11'
show_error_codes = true
strict_concatenate = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Meryll Dindin <merylldin@gmail.com>"]
description = "AWS routines for Python projects"
name = "sirtuin"
version = "0.1.0"

[tool.poetry.dependencies]
pydantic = "^1.10.2"
python = "3.11.*"
pyyaml = "^6.0"
tomli = "^2.0.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.982"
ruff = "^0.0.103"
toml-sort = "^0.20.1"
tomli = "^2.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[tool.poetry.group.types.dependencies]
types-pyyaml = "^6.0.12.1"

[tool.pytest.ini_options]
addopts = "--no-header -rA"
pythonpath = ["."]

[tool.ruff]
extend-select = ["Q"]
line-length = 88
target-version = "py311"
